Delete rome-osgi-test module
We're having issues building the project on Jenkins and releasing a new
version. The quick-fix is to delete the `rome-osgi-test` module, which
is not an ideal solution, but not super-bad either, because the module
is only used for testing and we rarely change anything osgi-related.

The errors from Jenkins:

```
Failed to transfer file:
http://repo1.maven.org/maven2/org/ops4j/pax/exam/pax-exam-inject/4.13.1/pax-exam-inject-4.13.1.jar.
Return code is: 501 , ReasonPhrase:HTTPS Required.
```

```
ERROR: Error creating bundle cache.
java.lang.Exception: Unable to lock bundle cache:
java.nio.channels.OverlappingFileLockException
```